+++ app/widgets/logger/src/Logger.kt
@@ -1,6 +1,8 @@
 package logger
 
+import javafx.geometry.Pos.TOP_LEFT
 import javafx.scene.control.TextArea
+import javafx.scene.layout.Priority.ALWAYS
 import mu.KLogging
 import sp.it.pl.layout.Widget
 import sp.it.pl.main.WidgetTags.DEVELOPMENT
@@ -9,17 +11,23 @@ import sp.it.pl.layout.WidgetCompanion
 import sp.it.pl.layout.controller.SimpleController
 import sp.it.pl.layout.feature.TextDisplayFeature
 import sp.it.pl.main.APP
-import sp.it.pl.main.IconUN
+import sp.it.pl.main.IconMA
+import sp.it.pl.main.IconOC
 import sp.it.pl.main.Widgets.LOGGER_NAME
 import sp.it.pl.main.emScaled
+import sp.it.pl.ui.objects.icon.CheckIcon
+import sp.it.pl.ui.objects.icon.Icon
 import sp.it.pl.ui.pane.ShortcutPane
 import sp.it.util.conf.cv
 import sp.it.util.conf.def
+import sp.it.util.conf.getDelegateConfig
 import sp.it.util.reactive.Subscribed
 import sp.it.util.reactive.consumeScrolling
 import sp.it.util.reactive.on
+import sp.it.util.ui.hBox
 import sp.it.util.ui.lay
 import sp.it.util.ui.prefSize
+import sp.it.util.ui.vBox
 import sp.it.util.ui.x
 import sp.it.util.units.version
 import sp.it.util.units.year
@@ -36,9 +44,15 @@ class Logger(widget: Widget): SimpleController(widget), TextDisplayFeature {
       root.prefSize = 500.emScaled x 500.emScaled
       root.consumeScrolling()
 
-      root.lay += area.apply {
-         isEditable = false
-         text = "# This is redirected output (System.out) stream of this application.\n"
+      root.lay += hBox(0.0, TOP_LEFT) {
+         lay += vBox {
+            lay += CheckIcon(wrapText).icons(IconMA.WRAP_TEXT).tooltip(::wrapText.getDelegateConfig().nameUi)
+            lay += Icon(IconOC.TRASHCAN).tooltip("Clear").onClickDo { area.clear() }
+         }
+         lay(ALWAYS) += area.apply {
+            isEditable = false
+            text = "# This is redirected output (System.out) stream of this application.\n"
+         }
       }
 
       stdoutReader.subscribe()
@@ -54,7 +68,7 @@ class Logger(widget: Widget): SimpleController(widget), TextDisplayFeature {
       override val name = LOGGER_NAME
       override val description = "Displays text or application standard output (stdout), which contains application logging."
       override val descriptionLong = "$description."
-      override val icon = IconUN(0x2e2a)
+      override val icon = IconOC.TERMINAL
       override val version = version(1, 1, 0)
       override val isSupported = true
       override val year = year(2015)